[fix] websocket 関連のコードを分けた
