+++ .env_template
@@ -1,3 +1,4 @@
 BUGSNAG_API_KEY = ""
 FLASK_ENV = "developement"
 DATABASE_URL = ""
+REDIS_URL = ""
