+++ api/_socket.py
@@ -1,3 +1,4 @@
 from flask_sockets import Sockets 
+from ._app import app
 
-sockets = Sockets()
+sockets = Sockets(app)
