+++ api/views.py
@@ -1,66 +1,12 @@
 from flask import Blueprint, make_response, jsonify, request, abort
 from api.models import Block, Map
 from api._db import db
-from api import sockets
-import gevent
-import redis
+from api._redis import redis_connection
 import json
-import threading
 
-redis = redis.from_url('redis://localhost:6379')
 
 
 api_app = Blueprint('api_app', __name__)
-class SocketClients():
-    def __init__(self):
-        self.clients = {}
-        self.pubsub = redis.pubsub()
-        self.pubsub.subscribe('received_message')
-
-    def __iter_data(self):
-        for data in self.pubsub.listen():
-            try:
-                json.loads(data.get('data'))
-            except:
-                continue
-            yield json.loads(data.get('data'))
-
-    def register(self, client, map_id):
-        if map_id in self.clients:
-            self.clients[map_id].append(client)
-        else:
-            self.clients[map_id] = [client]
-
-    def send(self, client, data, map_id):
-        try:
-            client.send(str(data))
-        except Exception:
-            self.clients[map_id].remove(client)
-
-    def run(self):
-        for data in self.__iter_data():
-            message = data.get('message')
-            map_id = data.get('map_id')
-            if data and map_id:
-                gevent.spawn(self.castForMap, message, map_id)
-
-    def start(self):
-        gevent.spawn(self.run)
-
-    def castForMap(self, message, map_id):
-        if map_id in self.clients:
-            for client in self.clients[map_id]:
-                gevent.spawn(self.send, client, message, map_id)
-
-    
-socket_clients = SocketClients()
-socket_clients.start()
-
-@sockets.route('/receive/<uuid:map_id>/')
-def outbox(ws, map_id):
-    socket_clients.register(ws, str(map_id))
-    while not ws.closed:
-        gevent.sleep(0.1)
 
 @api_app.route('/add_blocks/<uuid:map_id>/', methods=["POST"])
 def add_block(map_id):
@@ -73,7 +19,7 @@ def add_block(map_id):
         'message': request.json,
         'map_id': str(map_id)
     }
-    redis.publish('received_message', json.dumps(data)) 
+    redis_connection.publish('received_message', json.dumps(data))
     return make_response('ok')
