+++ settings.py
@@ -10,3 +10,4 @@ load_dotenv(dotenv_path=dotenv_path)
 BUGSNAG_API_KEY = os.environ.get("BUGSNAG_API_KEY")
 FLASK_ENV = os.environ.get("FLASK_ENV")
 DATABASE_URL = os.environ.get("DATABASE_URL")
+REDIS_URL = os.environ.get("REDIS_URL")
